+++ app/controllers/search_controller.rb
@@ -1,0 +1,11 @@
+class SearchController < ApplicationController
+
+  def new
+    binding.pry
+  end
+
+  def show
+    binding.pry
+    @country = CountryObject.find_by_country_code(params["Search"])
+  end
+end
