+++ app/views/search/new.html.erb
@@ -1,0 +1,34 @@
+
+<h1>Search the World Bank API by Country Code</h1>
+<div >
+  <%= form_tag('/', method: :get) do %>
+    <%= search_field_tag 'Search' %>
+    <%= submit_tag "Search" %>
+  <% end %>
+</div>
+
+<div class>
+  <table>
+    <caption>Country Information</caption>
+    <thead>
+      <tr>
+        <th>Name</th>
+        <th>Region</th>
+        <th>Income Level</th>
+        <th>Capital</th>
+        <th>Longitude</th>
+        <th>Latitude</th>
+      </tr>
+    </thead>
+    <tbody>
+      <tr>
+        <td><%= @country.name %></td>
+        <td><%= @country.region %></td>
+        <td><%= @country.income_level %></td>
+        <td><%= @country.capital %></td>
+        <td><%= @country.longitude %></td>
+        <td><%= @country.latitude %></td>
+      </tr>
+    </tbody>
+  </table>
+</div>
