+++ config/routes.rb
@@ -1,3 +1,6 @@
 Rails.application.routes.draw do
-  root to: 'regions#index', as: 'root'
+  root to: 'home#index', as: 'root'
+
+  # get 'search', to: 'search#show'
+  # root to: 'regions#index', as: 'root'
 end
